Finish initial version and add detail to README
Finished the last bit of code. Made the win conditions output a
winner in the console. Also cleaned up the getPlayerChoice function
in order to remove redundant commands and prevent the unneccesary
first prompt from showing up (it outputted nothing to the console).
Also added a little more detail about the project on the README.